### 2024-02-14 #6353 @editorialbot
[PRE REVIEW]: Monte Carlo / Dynamic Code (MC/DC): An accelerated Python package for fully transient neutron transport and rapid methods development
**Submitting author:** @jpmorgan98 (<a href="http://orcid.org/0000-0003-1379-5431">J. Piper Morgan</a>)
**Repository:** https://github.com/CEMeNT-PSAAP/MCDC
**Branch with paper.md** (empty if default branch):
**Version:** v0.9.0
