### 2024-02-14 4fe4ec4
adding caching functions for loop.py and adding args
