### 2024-02-14 4b9dcf7
adding documentation for cacheing and debug options
