+++ mcdc/loop.py
@@ -28,6 +28,7 @@ from mcdc.print_ import (
 #     to effectivly disable cache, delete the cache folder (often located in /MCDC/mcdc/__pycache__)
 #     see more about cacheing here https://numba.readthedocs.io/en/stable/developer/caching.html
 
+
 @njit(cache=True)
 def loop_fixed_source(mcdc):
     # Loop over batches
