### 2024-02-16 #6353 @lewisfish
> hi @lewisfish @csadorf 👋 would you be interested in and available for reviewing this JOSS submission?

Hi @kellyrowland, I'd be happy to review this. I'd note I'm not an expert in neutron transport, rather photon transport though they share a lot of the same basic physics and technqiues. If you're happy with this I'll start reviewing.
