### 2024-02-20 703a00d
Refactor checkpoint - loop source mostly resolved
