+++ mcdc/gpu_loop.py
@@ -1,6 +1,6 @@
 import numpy as np
 from numpy import ascontiguousarray as cga
-from numba import njit, objmode, jit
+from numba import njit, objmode, jit, cuda
 from scipy.linalg import eig
 
 import mcdc.kernel as kernel
@@ -10,7 +10,6 @@ import mcdc.print_ as print_module
 
 from mcdc.constant import *
 from mcdc.print_ import (
-    print_header_batch,
     print_progress,
     print_progress_eigenvalue,
     print_progress_iqmc,
@@ -18,112 +17,207 @@ from mcdc.print_ import (
     print_iqmc_eigenvalue_exit_code,
 )
 
-
-from mcdc.adapt import toggle, for_gpu, for_cpu, universal_arrays
 import mcdc.adapt as adapt
+from mcdc.adapt import toggle, for_gpu, for_cpu, universal_arrays
+
+import numba as nb
 
 # =========================================================================
-# Fixed-source loop
-# =========================================================================
-
-
-@njit
-def loop_fixed_source(mcdc):
-    # Loop over batches
-    for idx_batch in range(mcdc["setting"]["N_batch"]):
-        mcdc["idx_batch"] = idx_batch
-        seed_batch = kernel.split_seed(idx_batch, mcdc["setting"]["rng_seed"])
-
-        # Print multi-batch header
-        if mcdc["setting"]["N_batch"] > 1:
-            with objmode():
-                print_header_batch(mcdc)
-            if mcdc["technique"]["uq"]:
-                seed_uq = kernel.split_seed(seed_batch, SEED_SPLIT_UQ)
-                kernel.uq_reset(mcdc, seed_uq)
-
-        # Loop over time censuses
-        for idx_census in range(mcdc["setting"]["N_census"]):
-            mcdc["idx_census"] = idx_census
-            seed_census = kernel.split_seed(seed_batch, SEED_SPLIT_CENSUS)
-
-            # Loop over source particles
-            seed_source = kernel.split_seed(seed_census, SEED_SPLIT_SOURCE)
-            loop_source(seed_source, mcdc)
-
-            # Loop over source precursors
-            if mcdc["bank_precursor"]["size"] > 0:
-                seed_source_precursor = kernel.split_seed(
-                    seed_census, SEED_SPLIT_SOURCE_PRECURSOR
-                )
-                loop_source_precursor(seed_source_precursor, mcdc)
-
-            # Time census closeout
-            if idx_census < mcdc["setting"]["N_census"] - 1:
-                # TODO: Output tally (optional)
-
-                # Manage particle banks: population control and work rebalance
-                seed_bank = kernel.split_seed(seed_census, SEED_SPLIT_BANK)
-                kernel.manage_particle_banks(seed_bank, mcdc)
-
-        # Multi-batch closeout
-        if mcdc["setting"]["N_batch"] > 1:
-            # Reset banks
-            mcdc["bank_source"]["size"] = 0
-            mcdc["bank_census"]["size"] = 0
-            mcdc["bank_active"]["size"] = 0
-
-            # Tally history closeout
-            kernel.tally_reduce_bin(mcdc)
-            kernel.tally_closeout_history(mcdc)
-            # Uq closeout
-            if mcdc["technique"]["uq"]:
-                kernel.uq_tally_closeout_batch(mcdc)
-
-    # Tally closeout
-    if mcdc["technique"]["uq"]:
-        kernel.uq_tally_closeout(mcdc)
-    else:
-        kernel.tally_closeout(mcdc)
-
-
-# =========================================================================
-# Eigenvalue loop
+# Main loop
[…]
