+++ examples/fixed_source/kobayashi3-TD/input.py
@@ -68,7 +68,7 @@ mcdc.tally(
 )
 
 # Setting
-mcdc.setting(N_particle=1e6)
+mcdc.setting(N_particle=1e3)#6)
 mcdc.implicit_capture()
 
 # Run
