+++ mcdc/loop.py
@@ -131,83 +131,6 @@ def loop_eigenvalue(mcdc):
 # =============================================================================
 
 
-@njit
-def _loop_source(seed, mcdc):
-    ## Progress bar indicator
-    #N_prog = 0
-    #
-    # Loop over particle sources
-    #work_start = mcdc["mpi_work_start"]
-    #work_size = mcdc["mpi_work_size"]
-    #work_end = work_start + work_size
-    #for idx_work in range(work_size):
-    #    seed_work = kernel.split_seed(work_start + idx_work, seed)
-
-        ## Particle tracker
-        #if mcdc["setting"]["track_particle"]:
-        #    mcdc["particle_track_history_ID"] += 1
-
-        ## =====================================================================
-        ## Get a source particle and put into active bank
-        ## =====================================================================
-
-        ## Get from fixed-source?
-        #if mcdc["bank_source"]["size"] == 0:
-        #    # Sample source
-        #    P = kernel.source_particle(seed_work, mcdc)
-
-        ## Get from source bank
-        #else:
-        #    P = mcdc["bank_source"]["particles"][idx_work]
-
-        ## Check if it is beyond current census index
-        #idx_census = mcdc["idx_census"]
-        #if P["t"] > mcdc["setting"]["census_time"][idx_census]:
-        #    kernel.add_particle(P, mcdc["bank_census"])
-        #else:
-        #    # Add the source particle into the active bank
-        #    kernel.add_particle(P, mcdc["bank_active"])
-
-        # =====================================================================
-        # Run the source particle and its secondaries
-        # =====================================================================
-
-        # Loop until active bank is exhausted
-        #while mcdc["bank_active"]["size"] > 0:
-            # Get particle from active bank
-            #kernel.get_particle(P,mcdc["bank_active"], mcdc)
-
-            # Apply weight window
-            #if mcdc["technique"]["weight_window"]:
-            #    kernel.weight_window(P, mcdc)
-
-            ## Particle tracker
-            #if mcdc["setting"]["track_particle"]:
-            #    mcdc["particle_track_particle_ID"] += 1
-
-            # Particle loop
-            #loop_particle(P, mcdc)
-
-        ## =====================================================================
-        ## Closeout
-        ## =====================================================================
-
-        ## Tally history closeout for one-batch fixed-source simulation
-        #if not mcdc["setting"]["mode_eigenvalue"] and mcdc["setting"]["N_batch"] == 1:
-        #    kernel.tally_closeout_history(mcdc)
-
-        ## Tally history closeout for multi-batch uq simulation
-        #if mcdc["technique"]["uq"]:
-        #    kernel.uq_tally_closeout_history(mcdc)
-
-        ## Progress printout
-        #percent = (idx_work + 1.0) / mcdc["mpi_work_size"]
-        #if mcdc["setting"]["progress_bar"] and int(percent * 100.0) > N_prog:
-        #    N_prog += 1
-        #    with objmode():
-        #        print_progress(percent, mcdc)
-
-
 #$$$$GOOD
 @njit
 def generate_source_particle(idx_work, seed, prog):
