### 2024-02-21 4fc5544
debug install.sh and expand python requirement
