### 2024-02-21 3b4d6a3
revert back to previous python reqs
