+++ mcdc/loop.py
@@ -69,9 +69,12 @@ def loop_fixed_source(mcdc):
         # Multi-batch closeout
         if mcdc["setting"]["N_batch"] > 1:
             # Reset banks
-            mcdc["bank_source"]["size"] = 0
-            mcdc["bank_census"]["size"] = 0
-            mcdc["bank_active"]["size"] = 0
+            kernel.set_bank_size(mcdc["bank_source"],0)
+            kernel.set_bank_size(mcdc["bank_census"],0)
+            kernel.set_bank_size(mcdc["bank_active"],0)
+            #mcdc["bank_source"]["size"] = 0
+            #mcdc["bank_census"]["size"] = 0
+            #mcdc["bank_active"]["size"] = 0
 
             # Tally history closeout
             kernel.tally_reduce_bin(mcdc)
@@ -136,7 +139,7 @@ def loop_eigenvalue(mcdc):
 def generate_source_particle(idx_work, seed, prog):
     mcdc = adapt.device(prog)
 
-    seed_work = kernel.split_seed(idx_work, seed)
+    seed_work = kernel.split_seed(idx_work,seed)
 
     # =====================================================================
     # Get a source particle and put into active bank
@@ -145,12 +148,12 @@ def generate_source_particle(idx_work, seed, prog):
     # Get from fixed-source?
     if kernel.get_bank_size(mcdc["bank_source"]) == 0:
         # Sample source
-        P : type_.particle_record = kernel.source_particle(seed_work, mcdc)
+        P : type_.particle_record = kernel.source_particle(seed_work,mcdc)
 
     # Get from source bank
     else:
         P : type_.particle_record = mcdc["bank_source"]["particles"][idx_work]
-    
+
     # Particle tracker
     if mcdc["setting"]["track_particle"]:
         kernel.allocate_hid(P,mcdc)
@@ -177,7 +180,7 @@ def prep_particle(P,prog):
     # Particle tracker
     if mcdc["setting"]["track_particle"]:
         kernel.allocate_pid(P,mcdc)
-    
+
     # Particle tracker
     if mcdc["setting"]["track_particle"]:
         kernel.track_particle(P, mcdc)
@@ -208,7 +211,7 @@ def process_sources(seed,mcdc):
     work_size = mcdc["mpi_work_size"]
     work_end = work_start + work_size
     for idx_work in range(work_size):
-        seed_work = kernel.split_seed(work_start + idx_work, seed)
+        #seed_work = kernel.split_seed(work_start + idx_work, seed)
 
         generate_source_particle(idx_work,seed,mcdc)
 
@@ -731,25 +734,245 @@ def davidson(mcdc):
 # =============================================================================
 
 
-@njit
-def loop_source_precursor(seed, mcdc):
+#@njit
+#def loop_source_precursor(seed, mcdc):
     # TODO: censussed neutrons seeding is still not reproducible
 
     # Progress bar indicator
-    N_prog = 0
+    #N_prog = 0
 
     # =========================================================================
     # Sync. RNG skip ahead for reproducibility
     # =========================================================================
 
     # Exscan upper estimate of number of particles generated locally
-    idx_start, N_local, N_global = kernel.bank_scanning_DNP(
-        mcdc["bank_precursor"], mcdc
-    )
+    #idx_start, N_local, N_global = kernel.bank_scanning_DNP(
+    #    mcdc["bank_precursor"], mcdc
+    #)
 
     # =========================================================================
     # Loop over precursor sources
     # =========================================================================
+
+    #for idx_work in range(mcdc["mpi_work_size_precursor"]):
+        ## Get precursor
+        #DNP = mcdc["bank_precursor"]["precursors"][idx_work]
+
+        ## Set groups
+        #j = DNP["g"]
+        #g = DNP["n_g"]
+
[…]
