### 2024-02-22 ad52627
Checkpoint commit. Kernel launches, but crashes.
