+++ mcdc/kernel.py
@@ -1855,7 +1855,7 @@ def move_to_event(P, mcdc):
             track_particle(P, mcdc)
         iqmc_score_tallies(P, distance, mcdc)
         iqmc_continuous_weight_reduction(P, distance, mcdc)
-        if np.abs(P["w"]) <= mcdc["technique"]["iqmc"]["w_min"]:
+        if math.fabs(P["w"]) <= mcdc["technique"]["iqmc"]["w_min"]:
             P["alive"] = False
 
     # Score tracklength tallies
@@ -2063,7 +2063,7 @@ def surface_crossing(P, prog):
         material_ID_new = get_particle_material(P, mcdc)
         if material_ID_old != material_ID_new:
             # Sample derivative source particles
-            sensitivity_surface(P, surface, material_ID_old, material_ID_new, mcdc)
+            sensitivity_surface(P, surface, material_ID_old, material_ID_new, prog)
 
 
 # =============================================================================
@@ -2568,7 +2568,7 @@ def fission_CE(P, nuclide, P_new):
     J = 6
     nu = get_nu(NU_FISSION, nuclide, E)
     nu_p = get_nu(NU_FISSION_PROMPT, nuclide, E)
-    nu_d = np.zeros(J)
+    nu_d = adapt.local_j_array()["values"]#np.zeros(J)
     for j in range(J):
         nu_d[j] = get_nu_group(NU_FISSION_DELAYED, nuclide, E, j)
 
@@ -2717,7 +2717,7 @@ def weight_window(P, prog):
 # ==============================================================================
 
 
-@njit
+@toggle("iQMC")
 def iqmc_continuous_weight_reduction(P, distance, mcdc):
     """
     Continuous weight reduction technique based on particle track-length, for
@@ -2747,7 +2747,7 @@ def iqmc_continuous_weight_reduction(P, distance, mcdc):
     P["w"] = iqmc_w_sum
 
 
-@njit
+@toggle("iQMC")
 def iqmc_preprocess(mcdc):
     # set bank source
     iqmc = mcdc["technique"]["iqmc"]
@@ -2764,7 +2764,7 @@ def iqmc_preprocess(mcdc):
     iqmc_consolidate_sources(mcdc)
 
 
-@njit
+@toggle("iQMC")
 def iqmc_prepare_nuSigmaF(mcdc):
     iqmc = mcdc["technique"]["iqmc"]
     mesh = iqmc["mesh"]
@@ -2786,7 +2786,7 @@ def iqmc_prepare_nuSigmaF(mcdc):
                     )
 
 
-@njit
+@toggle("iQMC")
 def iqmc_prepare_source(mcdc):
     """
     Iterates trhough all spatial cells to calculate the iQMC source. The source
@@ -2826,7 +2826,7 @@ def iqmc_prepare_source(mcdc):
     iqmc_update_source(mcdc)
 
 
-@njit
+@toggle("iQMC")
 def iqmc_prepare_particles(mcdc):
     """
     Create N_particles assigning the position, direction, and group from the
@@ -2884,7 +2884,7 @@ def iqmc_prepare_particles(mcdc):
         adapt.add_source(P_new, mcdc)
 
 
-@njit
+@toggle("iQMC")
 def iqmc_res(flux_new, flux_old):
     """
 
@@ -2909,7 +2909,7 @@ def iqmc_res(flux_new, flux_old):
     return (flux_new - flux_old) / flux_old
 
 
-@njit
+@toggle("iQMC")
[…]
