+++ mcdc/loop.py
@@ -1292,7 +1292,7 @@ def make_gpu_process_sources(precursor):
             mcdc["bank_census"] = mcdc_new["bank_census"]
             mcdc["bank_source"] = mcdc_new["bank_source"]
             mcdc["bank_precursor"] = mcdc_new["bank_precursor"]
-            mcdc["i_cycle"] = mcdc_new["i_cycle"]
+            #mcdc["i_cycle"] = mcdc_new["i_cycle"]
             mcdc["cycle_active"] = mcdc_new["cycle_active"]
             mcdc["eigenvalue_tally_nuSigmaF"] = mcdc_new["eigenvalue_tally_nuSigmaF"]
             mcdc["eigenvalue_tally_n"] = mcdc_new["eigenvalue_tally_n"]
@@ -1325,7 +1325,7 @@ def make_gpu_process_sources(precursor):
             mcdc["bank_census"] = mcdc_new["bank_census"]
             mcdc["bank_source"] = mcdc_new["bank_source"]
             mcdc["bank_precursor"] = mcdc_new["bank_precursor"]
-            mcdc["i_cycle"] = mcdc_new["i_cycle"]
+            #mcdc["i_cycle"] = mcdc_new["i_cycle"]
             mcdc["cycle_active"] = mcdc_new["cycle_active"]
             mcdc["eigenvalue_tally_nuSigmaF"] = mcdc_new["eigenvalue_tally_nuSigmaF"]
             mcdc["eigenvalue_tally_n"] = mcdc_new["eigenvalue_tally_n"]
