+++ mcdc/type_.py
@@ -56,8 +56,6 @@ def align(field_list):
         if len(field) == 3:
             field = ( field[0], field[1], fixup_dims(field[2]) )
             for d in field[2]:
-                if d == 0:
-                    print("AAAAAAAAAAAAAAAAAA")
                 multiplier *= d
         kind = np.dtype(field[1])
         size = kind.itemsize
@@ -89,7 +87,7 @@ def align(field_list):
         result.append((f"padding_{pad_id}", uint8, (pad_size,)))
         pad_id += 1
 
-    print(result)
+    #print(result)
     return result
 
 
@@ -199,8 +197,8 @@ def make_type_particle_record(input_deck):
         ("E", float64),
         ("w", float64),
         ("sensitivity_ID", int64),
+        ("alive", bool_), # Should not be last until padding before nested structs is figured out
         ("rng_seed", uint64),
-        ("alive", bool_),
     ]
 
     # Get modes
@@ -219,10 +217,12 @@ def make_type_particle_record(input_deck):
     iqmc_struct = [("w", float64, (G,))]
     if input_deck.setting["track_particle"]:
         struct += [("track_pid",int64),("track_hid",int64)]
-    struct += [("iqmc", iqmc_struct)]
+    struct += [("iqmc", into_dtype(iqmc_struct))]
+    print(np.dtype(struct))
 
     # Save type
     particle_record = into_dtype(struct)
+    print(particle_record)
 
 
 precursor = into_dtype(
@@ -244,13 +244,13 @@ precursor = into_dtype(
 
 def particle_bank(max_size):
     return into_dtype(
-        [("particles", particle_record, (max_size,)), ("size", int64, (1,)), ("tag", "U16")]
+        [("particles", particle_record, (max_size,)), ("size", int64, (1,)), ("tag", "U32")]
     )
 
 
 def precursor_bank(max_size):
     return into_dtype(
-        [("precursors", precursor, (max_size,)), ("size", int64, (1,)), ("tag", "U16")]
+        [("precursors", precursor, (max_size,)), ("size", int64, (1,)), ("tag", "U32")]
     )
 
 
@@ -727,7 +727,7 @@ def make_type_setting(deck):
         ("mode_CE", bool_),
         # Misc.
         ("progress_bar", bool_),
-        ("output_name", "U30"),
+        ("output_name", str_),
         ("save_input_deck", bool_),
         ("track_particle", bool_),
         # Eigenvalue mode
@@ -744,10 +744,10 @@ def make_type_setting(deck):
         ("census_time", float64, (card["N_census"],)),
         # Particle source file
         ("source_file", bool_),
-        ("source_file_name", "U30"),
+        ("source_file_name", str_),
         # Initial condition source file
         ("IC_file", bool_),
-        ("IC_file_name", "U30"),
+        ("IC_file_name", str_),
         ("N_precursor", uint64),
         # TODO: Move to technique
         ("N_sensitivity", uint64),
@@ -1249,8 +1249,9 @@ param_names = ["tag", "ID", "key", "mean", "delta", "distribution", "rng_seed"]
 # Global
 # ==============================================================================
 
-def copy_global(dst,src):
-    pass
+
+
+
 
 def make_type_global(input_deck):
     global global_
@@ -1324,6 +1325,7 @@ def make_type_global(input_deck):
     # GLobal type
     global_ = into_dtype(
         [
+            ("mpi_work_iter", int64, (1,)),
             ("nuclides", nuclide, (N_nuclide,)),
             ("materials", material, (N_material,)),
             ("surfaces", surface, (N_surface,)),
@@ -1352,7 +1354,6 @@ def make_type_global(input_deck):
[…]
