### 2024-02-27 #6353 @matt-graham
> hi szabo137 matt-graham 👋 would either of you be interested in and available to review this JOSS submission? it looks like you're both participating in ongoing reviews, so no worries if you don't have the bandwidth.

Hi @kellyrowland. Realise you've now got reviewers sorted, but if you do end up needing a further reviewer I'd be happy to review this but would not have capacity till after mid-March.
