### 2024-02-27 9c4a651
Added python version check to install script
