+++ mcdc/loop.py
@@ -287,18 +287,18 @@ def loop_source_dd(seed, mcdc):
             kernel.tally_closeout_history(mcdc)
 
         # Progress printout
-        percent = ((work_idx + 1.0) / mcdc["mpi_work_size"])*0.5
+        percent = ((work_idx + 1.0) / mcdc["mpi_work_size"]) * 0.5
         if mcdc["setting"]["progress_bar"] and int(percent * 100.0) > N_prog:
             N_prog += 1
             with objmode():
-                print_progress_dd(percent, mcdc,"sourcing")
+                print_progress_dd(percent, mcdc, "sourcing")
 
     kernel.dd_particle_send(mcdc)
     terminated = False
     kernel.dd_particle_receive(mcdc)
     wr_new = 0
     max_work = 1
-    #print_msg("Done sourcing, now running")
+    # print_msg("Done sourcing, now running")
     while not terminated:
         if mcdc["bank_active"]["size"] > 0:
             # wr_new = 0
@@ -338,12 +338,12 @@ def loop_source_dd(seed, mcdc):
             max_work = work_remaining
 
         # Progress printout
-        percent = (1-work_remaining/max_work)*0.5+0.5
+        percent = (1 - work_remaining / max_work) * 0.5 + 0.5
         if mcdc["setting"]["progress_bar"] and int(percent * 100.0) > N_prog:
             N_prog += 1
             with objmode():
-                print_progress_dd(percent, mcdc,"running")
-        #print_msg(str(percent)+"percent"+ str(work_remaining)+','+str(max_work))
+                print_progress_dd(percent, mcdc, "running")
+        # print_msg(str(percent)+"percent"+ str(work_remaining)+','+str(max_work))
         if work_remaining == 0:
             wr_new += 1
         else:
