+++ mcdc/print_.py
@@ -90,13 +90,16 @@ def print_progress(percent, mcdc):
                 )
         sys.stdout.flush()
 
-def print_progress_dd(percent, mcdc,stage):
+
+def print_progress_dd(percent, mcdc, stage):
     if master:
         sys.stdout.write("\r")
         if not mcdc["setting"]["mode_eigenvalue"]:
             if mcdc["setting"]["N_census"] == 1:
                 sys.stdout.write(
-                    "DD:" + stage + " [%-28s] %d%%" % ("=" * int(percent * 28), percent * 100.0)
+                    "DD:"
+                    + stage
+                    + " [%-28s] %d%%" % ("=" * int(percent * 28), percent * 100.0)
                 )
             else:
                 idx = mcdc["idx_census"] + 1
