### 2024-03-16 86410bd
Improved termination criteria
