### 2024-03-16 a96a3c5
Improved termination criteria
