+++ mcdc/main.py
@@ -570,7 +570,7 @@ def prepare():
     mcdc["technique"]["wr_threshold"] = input_deck.technique["wr_threshold"]
 
     # Survival probability
-    # mcdc["technique"]["wr_chance"] = input_deck.technique["wr_chance"]
+    mcdc["technique"]["wr_survive"] = input_deck.technique["wr_survive"]
     # =========================================================================
     # Domain Decomposition
     # =========================================================================
