+++ mcdc/type_.py
@@ -694,7 +694,7 @@ def make_type_technique(input_deck):
         ("iQMC", bool_),
         ("IC_generator", bool_),
         ("branchless_collision", bool_),
-        ("domain_decomp", bool_),
+        ("domain_decomposition", bool_),
         ("uq", bool_),
     ]
 
@@ -708,19 +708,19 @@ def make_type_technique(input_deck):
     # domain decomp
     # =========================================================================
     # Mesh
-    mesh, Nx, Ny, Nz, Nt, Nmu, N_azi, Ng = make_type_mesh(card["domain_mesh"])
-    struct += [("domain_mesh", mesh)]
-    struct += [("d_idx", int64)]
-    struct += [("sent", int64)]
-    struct += [("work_ratio", int64, (len(card["work_ratio"]),))]
-    struct += [("exchange_rate", int64)]
-    struct += [("repro", bool_)]
-    struct += [("xp_neigh", int64, (len(card["xp_neigh"]),))]
-    struct += [("xn_neigh", int64, (len(card["xn_neigh"]),))]
-    struct += [("yp_neigh", int64, (len(card["yp_neigh"]),))]
-    struct += [("yn_neigh", int64, (len(card["yn_neigh"]),))]
-    struct += [("zp_neigh", int64, (len(card["zp_neigh"]),))]
-    struct += [("zn_neigh", int64, (len(card["zn_neigh"]),))]
+    mesh, Nx, Ny, Nz, Nt, Nmu, N_azi, Ng = make_type_mesh(card["dd_mesh"])
+    struct += [("dd_mesh", mesh)]
+    struct += [("dd_idx", int64)]
+    struct += [("dd_sent", int64)]
+    struct += [("dd_work_ratio", int64, (len(card["dd_work_ratio"]),))]
+    struct += [("dd_exchange_rate", int64)]
+    struct += [("dd_repro", bool_)]
+    struct += [("dd_xp_neigh", int64, (len(card["dd_xp_neigh"]),))]
+    struct += [("dd_xn_neigh", int64, (len(card["dd_xn_neigh"]),))]
+    struct += [("dd_yp_neigh", int64, (len(card["dd_yp_neigh"]),))]
+    struct += [("dd_yn_neigh", int64, (len(card["dd_yn_neigh"]),))]
+    struct += [("dd_zp_neigh", int64, (len(card["dd_zp_neigh"]),))]
+    struct += [("dd_zn_neigh", int64, (len(card["dd_zn_neigh"]),))]
 
     # =========================================================================
     # Weight window
@@ -1076,7 +1076,7 @@ def make_type_global(input_deck):
     bank_precursor = precursor_bank(0)
 
     # Domain banks if needed
-    if input_deck.technique["domain_decomp"]:
+    if input_deck.technique["domain_decomposition"]:
         bank_domain_xp = particle_bank(input_deck.technique["domain_bank_size"])
         bank_domain_xn = particle_bank(input_deck.technique["domain_bank_size"])
         bank_domain_yp = particle_bank(input_deck.technique["domain_bank_size"])
@@ -1138,7 +1138,7 @@ def make_type_global(input_deck):
             ("bank_domain_zp", bank_domain_zp),
             ("bank_domain_zn", bank_domain_zn),
             ("bank_precursor", bank_precursor),
-            ("d_idx", int64),
+            ("dd_idx", int64),
             ("k_eff", float64),
             ("k_cycle", float64, (N_cycle,)),
             ("k_avg", float64),
