+++ mcdc/card.py
@@ -80,12 +80,12 @@ class InputDeck:
             "ww": np.ones([1, 1, 1, 1]),
             "ww_width": 2.5,
             "ww_mesh": make_card_mesh(),
-            "domain_decomp": False,
-            "d_idx": 0,
-            "domain_mesh": make_card_mesh(),
-            "exchange_rate": 0,
-            "repro": False,
-            "work_ratio": np.array([1]),
+            "domain_decomposition": False,
+            "dd_idx": 0,
+            "dd_mesh": make_card_mesh(),
+            "dd_exchange_rate": 0,
+            "dd_repro": False,
+            "dd_work_ratio": np.array([1]),
             "weight_roulette": False,
             "wr_threshold": 0.0,
             "wr_survive": 1.0,
