+++ mcdc/kernel.py
@@ -24,9 +24,9 @@ from mcdc.loop import loop_source
 @njit
 def domain_crossing(P, mcdc):
     # Domain mesh crossing
-    max_size = mcdc["technique"]["exchange_rate"]
-    if mcdc["technique"]["domain_decomp"]:
-        mesh = mcdc["technique"]["domain_mesh"]
+    max_size = mcdc["technique"]["dd_exchange_rate"]
+    if mcdc["technique"]["domain_decomposition"]:
+        mesh = mcdc["technique"]["dd_mesh"]
         # Determine which dimension is crossed
         x, y, z, t, directions = mesh_crossing_evaluate(P, mesh)
         if len(directions) == 0:
@@ -78,84 +78,84 @@ def dd_particle_send(mcdc):
     with objmode():
         for i in range(
             max(
-                len(mcdc["technique"]["xp_neigh"]),
-                len(mcdc["technique"]["xn_neigh"]),
-                len(mcdc["technique"]["yp_neigh"]),
-                len(mcdc["technique"]["yn_neigh"]),
-                len(mcdc["technique"]["zp_neigh"]),
-                len(mcdc["technique"]["zn_neigh"]),
+                len(mcdc["technique"]["dd_xp_neigh"]),
+                len(mcdc["technique"]["dd_xn_neigh"]),
+                len(mcdc["technique"]["dd_yp_neigh"]),
+                len(mcdc["technique"]["dd_yn_neigh"]),
+                len(mcdc["technique"]["dd_zp_neigh"]),
+                len(mcdc["technique"]["dd_zn_neigh"]),
             )
         ):
-            if mcdc["technique"]["xp_neigh"].size > i:
+            if mcdc["technique"]["dd_xp_neigh"].size > i:
                 size = mcdc["bank_domain_xp"]["size"]
-                ratio = int(size / len(mcdc["technique"]["xp_neigh"]))
+                ratio = int(size / len(mcdc["technique"]["dd_xp_neigh"]))
                 start = ratio * i
                 end = start + ratio
-                if i == len(mcdc["technique"]["xp_neigh"]) - 1:
+                if i == len(mcdc["technique"]["dd_xp_neigh"]) - 1:
                     end = size
                 bank = np.array(mcdc["bank_domain_xp"]["particles"][start:end])
                 request1 = MPI.COMM_WORLD.send(
-                    bank, dest=mcdc["technique"]["xp_neigh"][i], tag=1
+                    bank, dest=mcdc["technique"]["dd_xp_neigh"][i], tag=1
                 )
 
-            if mcdc["technique"]["xn_neigh"].size > i:
+            if mcdc["technique"]["dd_xn_neigh"].size > i:
                 size = mcdc["bank_domain_xn"]["size"]
-                ratio = int(size / len(mcdc["technique"]["xn_neigh"]))
+                ratio = int(size / len(mcdc["technique"]["dd_xn_neigh"]))
                 start = ratio * i
                 end = start + ratio
-                if i == len(mcdc["technique"]["xn_neigh"]) - 1:
+                if i == len(mcdc["technique"]["dd_xn_neigh"]) - 1:
                     end = size
                 bank = np.array(mcdc["bank_domain_xn"]["particles"][start:end])
                 request2 = MPI.COMM_WORLD.send(
-                    bank, dest=mcdc["technique"]["xn_neigh"][i], tag=2
[…]

[…]
