### 2024-03-18 3fcdd69
Consolidated domain decomposition into one fixed source loop (and removed unneeded prints)
