### 2024-03-18 0eb109b
Commented out dd progress print for numba compatability and removed old dd_loop
