### 2024-03-21 CEMeNT-PSAAP/MCDC#170 @lewisfish
Installation issue
Hi MCDC team,
I'm one of the reviewers working on reviewing your JOSS submission https://github.com/openjournals/joss-reviews/issues/6415.

Currently following the instructions in the README and the readthedocs webpage fails to install the package.
Currently on creating a Conda env, cloning the repo and running bash install.sh gives:
```
CommandNotFoundError: No command 'conda mpi4py'.
```
I see from PR #162  and #165  it has been fixed but not merged into the main branch.
Can this be merged into the main branch so I can tick off the installation item on my review list?
