### 2024-03-21 5312504
Buf fix: remove __pycache__ from names list in regresss tests if present
