### 2024-03-22 CEMeNT-PSAAP/MCDC#173 @lewisfish
Broken examples
Hi MCDC team,

I've found the following examples won't run or plot correctly:

- fixed_source/pulsed_sphere gives the following error:
```  File "/home/JOSS/MCDC/examples/fixed_source/pulsed_sphere/input.py", line 10, in <module>
    mat_iron = mcdc.material(
               ^^^^^^^^^^^^^^
  File "/home/JOSS/MCDC/mcdc/input_.py", line 326, in material
    with h5py.File(dir_name + "/" + nuc_name + ".h5", "r") as f:
                   ~~~~~~~~~^~~~~
TypeError: unsupported operand type(s) for +: 'NoneType' and 'str'
```
- fixed_source/inf_pin_ce has the same error as above.
- fixed_source/slab_ce has the same issue as above. This example could also benefit from more instructions (what is build_xml.py purpose?) and a mention that openmc is a dependency.
- fixed_source/slab_reed_iqmc is missing a plt.show() or savefig from the process.py file.
- eigenvalue/2d_c5g7 fails to plot with the following error. I think this is due to the "C" argument to pcolormesh only having 1 dimension
```
  File "/home/JOSS/MCDC/examples/eigenvalue/2d_c5g7/process.py", line 95, in <module>
    plt.pcolormesh(X, Y, phi_thermal_sd, shading="nearest")
  File "/home/anaconda3/envs/mcdc-env/lib/python3.11/site-packages/matplotlib/pyplot.py", line 3493, in pcolormesh
    __ret = gca().pcolormesh(
            ^^^^^^^^^^^^^^^^^
  File "/home/anaconda3/envs/mcdc-env/lib/python3.11/site-packages/matplotlib/__init__.py", line 1465, in inner
    return func(ax, *map(sanitize_sequence, args), **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/anaconda3/envs/mcdc-env/lib/python3.11/site-packages/matplotlib/axes/_axes.py", line 6292, in pcolormesh
    X, Y, C, shading = self._pcolorargs('pcolormesh', *args,
                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/anaconda3/envs/mcdc-env/lib/python3.11/site-packages/matplotlib/axes/_axes.py", line 5815, in _pcolorargs
    nrows, ncols = C.shape[:2]
    ^^^^^^^^^^^^
ValueError: not enough values to unpack (expected 2, got 1)
```
- eigenvalue/slab_kornreich_iqmc is missing a plt.show() or savefig from the process.py file.
- eigenvalue/smrg7 fails with ```ERROR: Particle census bank is full```
- c5g7/3d/TDX the default particle number does not resolve to the reference solution.

https://github.com/openjournals/joss-reviews/issues/6415
