### 2024-03-22 CEMeNT-PSAAP/MCDC#174 @lewisfish
Missing/incomplete Documentation
Hi MCDC team,

Some of my concerns on the documentation of the project.

Major points:

- The API is not fully documented. For example cell, universe, lattice, tally, eigenmode, setting, iQMC, are among some of the functions not documented. Please add doc strings for all user-space functions detailed in the docs. I would also highly suggest adding doc strings to internal functions as well.

Minor points:

- Following through the user guide, I've noticed that the reference solution isn't explictly given in the docs. It might be a good idea to add either an explicit link to this or as you've done already for the rest of the example, add the full code from the reference solution.
- It might be a good idea to match the MPI runnning instructions given in the readme with those in the docs. For example, adding the bit about mpiexec/mpirun and how to control the number of cores used.
- It might be due to me not being an expert in neutron transport, but the examples are a bit opaque to me. It would be helpful if there was a blurb or readme attached to each example explaning what experiment/setup is being simulated or even what parts of MCDC are being shown off. It might also be good to give the user a hint on whether to use Numba and/or MPI for each example as not to waste their time
- On line 303 of user.rst there is a typo "statical" should be statistical.

https://github.com/openjournals/joss-reviews/issues/6415
