### 2024-03-22 #6415 @lewisfish
@jpmorgan98 For the section in the paper discussing MCDC vs other codes. How does MCDC compare with open source neutron codes such as openMC or GEANT4? Also what codes are hard to aquire, install etc (guessing MCNP?)? It would be good to cite these packages (or others if the ones I chose are not appropriate).
