### 2024-03-22 #6415 @jpmorgan98
`OpenMC` is the other major open source transport code we usually make comparisons to but we can certainly talk about Geant4 as well. I will add that to the paper.

`MCNP` and `Shift` where the two major softwares we had in mind for difficult to access. Both of them are export controlled and only available via RSICC which can come with significant cost and time delays. I will expand on that in the paper and add those citations.

I will also dig into those issues with the team!

Thanks!
