### 2024-03-22 CEMeNT-PSAAP/MCDC#173 @jpmorgan98
The issue with the following examples:

- fixed_source/pulsed_sphere
- fixed_source/inf_pin_ce
- fixed_source/slab_ce

I believe is because the continuous energy data library is not configured. (@ilhamv please confirm). If this is the case due to legal requirements we cannot publish and distribute that data for use. I will add a runtime error that will stop the simulation in this case and add notes to the example files themselves. See [our docs](https://cement-psaapgithubio.readthedocs.io/en/latest/install.html#configuring-continuous-energy-library).
