### 2024-03-22 b704de0
adding plt.show
