+++ examples/eigenvalue/slab_kornreich_iqmc/process.py
@@ -95,3 +95,4 @@ plt.ylabel(r"$\phi(x)$")
 plt.xlabel(r"$x$")
 plt.grid()
 plt.legend()
+plt.show()
