### 2024-03-22 eb78022
adding plt.show
