+++ examples/fixed_source/slab_reed_iqmc/process.py
@@ -132,3 +132,4 @@ plt.ylabel(r"$\phi(x)$")
 plt.xlabel(r"$x$")
 plt.grid()
 plt.legend()
+plt.show()
