### 2024-03-22 5f4f697
adding error message for uncondifgured cont eng xsec data lib
