### 2024-03-23 bd0f78c
adding doc strings to all input functions
