+++ mcdc/input_.py
@@ -1,6 +1,7 @@
 """
 This module contains functions for setting MC/DC input deck.
 The input deck class is defined in `card.py` and instantiated in `global_.py`.
+Note docstring are numpy typed 
 """
 
 import h5py, math, mpi4py, os
@@ -656,6 +657,21 @@ def _set_J(x, t, card):
 
 
 def cell(surfaces_flags, fill, lattice_center=None):
+    """_summary_
+
+    Parameters
+    ----------
+    surfaces_flags : bool
+        list of surface cards
+    fill : bool
+        fill the cells
+    lattice_center : _type_, optional
+        _description_, by default None
+
+    Returns
+    -------
+    cell card : dictionary
+    """
     N_surface = len(surfaces_flags)
 
     # Make cell card
@@ -685,6 +701,20 @@ def cell(surfaces_flags, fill, lattice_center=None):
 
 
 def universe(cells, root=False):
+    """_summary_
+
+    Parameters
+    ----------
+    cells : dictionary
+        cell card dexcriptions
+    root : bool, optional
+        by default False
+
+    Returns
+    -------
+    dictionary
+        universe
+    """
     N_cell = len(cells)
 
     # Set default card values (c.f. type_.py)
@@ -708,6 +738,25 @@ def universe(cells, root=False):
 
 
 def lattice(x=None, y=None, z=None, universes=None):
+    """
+    Create a lattice card
+
+    Parameters
+    ----------
+    x : array like, optional
+        x[float], by default None
+    y : array like, optional
+        y[float], by default None
+    z : array like, optional
+        z[float], by default None
+    universes : dictionary, optional
+        d, by default None
+
+    Returns
+    -------
+    dictionary
+        lattice card
+    """
     # Make lattice card
     card = make_card_lattice()
     card["ID"] = len(mcdc.input_deck.lattices)
@@ -899,6 +948,35 @@ def tally(
     g=np.array([-INF, INF]),
     E=np.array([0.0, INF]),
 ):
+    """_summary_
+
+    Parameters
+    ----------
+    scores : _type_
+        _description_
+    x : _type_, optional
+        _description_, by default np.array([-INF, INF])
+    y : _type_, optional
+        _description_, by default np.array([-INF, INF])
+    z : _type_, optional
+        _description_, by default np.array([-INF, INF])
+    t : _type_, optional
+        _description_, by default np.array([-INF, INF])
+    mu : _type_, optional
+        _description_, by default np.array([-1.0, 1.0])
+    azi : _type_, optional
+        _description_, by default np.array([-PI, PI])
+    g : _type_, optional
+        _description_, by default np.array([-INF, INF])
+    E : _type_, optional
+        _description_, by default np.array([0.0, INF])
+
+    Returns
+    -------
+    _type_
+        _description_
+    """
+
     # Get tally card
     card = mcdc.input_deck.tally
 
@@ -940,6 +1018,44 @@ def tally(
 
 
 def setting(**kw):
+    """
+    Create a setting card.
+
+    Other Parameters
+    ----------------
+    N_particle : int
+        Number of MC particle histories to run (for k-eigen and iQMC its /iteration)
+    N_batch : int
+        Number of batches to run
+    rng_seed : int
+        Random number seed
+    time_boundary : float
+        the time edge of the problem after which all particles will be killed
+    progress_bar : bool
+        display the progress bar (disable when running MC/DC in a loop)
+    output_name : str
+        Name of the output file MC/DC should save data in, default is output.h5
+    save_input_deck : bool
+        toggle to save the input deck
+    particle_tracker : bool
+        true will track the paths of all individual particle histories, memory issues abound
+    k_eff : str
+        toggle true to run a k-eigen problem
+    source_file : str
+        specify a source file path and name
+    IC_file : str
[…]
