+++ docs/paper.bib
@@ -116,3 +116,84 @@
 }
 
 
+#shift
+@article{shift,
+	title = {Continuous-energy {Monte} {Carlo} neutron transport on {GPUs} in the {Shift} code},
+	volume = {128},
+	issn = {03064549},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/S0306454919300167},
+	doi = {10.1016/j.anucene.2019.01.012},
+	abstract = {A continuous-energy Monte Carlo neutron transport solver executing on GPUs has been developed within the Shift code. Several algorithmic approaches are considered, including both history-based and event-based implementations. Unlike in previous work involving multigroup Monte Carlo transport, it is demonstrated that event-based algorithms signiﬁcantly outperform a historybased approach for continuous-energy transport as a result of increased device occupancy and reduced thread divergence. Numerical results are presented for detailed full-core models of a small modular reactor (SMR), including a model containing depleted fuel materials. These results demonstrate the substantial gains in performance that are possible with the latest-generation of GPUs. On the depleted SMR core conﬁguration, an NVIDIA P100 GPU with 56 streaming multiprocessors provides performance equivalent to 90 CPU cores, and the latest V100 GPU with 80 multiprocessors oﬀers the performance of more than 150 CPU cores.},
+	language = {en},
+	urldate = {2023-11-03},
+	journal = {Annals of Nuclear Energy},
+	author = {Hamilton, Steven P. and Evans, Thomas M.},
+	month = jun,
+	year = {2019},
+	pages = {236--247},
+	file = {1-s2.0-S0306454919300167-am.pdf:/Users/jonesy/Documents/PapersLibrary/1-s2.0-S0306454919300167-am.pdf:application/pdf},
+}
+
+
+#mcatk
+@article{mcatk,
+	title = {Monte {Carlo} {Application} {ToolKit} ({MCATK})},
+	volume = {82},
+	issn = {0306-4549},
+	url = {https://www.sciencedirect.com/science/article/pii/S0306454914004472},
+	doi = {https://doi.org/10.1016/j.anucene.2014.08.047},
+	abstract = {The Monte Carlo Application ToolKit (MCATK) is a component-based software library designed to build specialized applications and to provide new functionality for existing general purpose Monte Carlo radiation transport codes. We will describe MCATK and its capabilities along with presenting some verification and validations results.},
+	journal = {Annals of Nuclear Energy},
+	author = {Adams, Terry and Nolen, Steve and Sweezy, Jeremy and Zukaitis, Anthony and Campbell, Joann and Goorley, Tim and Greene, Simon and Aulwes, Rob},
+	year = {2015},
+	keywords = {Agile development, Component software, Monte Carlo particle transport, Parallel computing, Population control, Time-dependent},
+	pages = {41--47},
+	annote = {Joint International Conference on Supercomputing in Nuclear Applications and Monte Carlo 2013, SNA + MC 2013. Pluri- and Trans-disciplinarity, Towards New Modeling and Numerical Simulation Paradigms},
+}
+
+
+# mcnp
+@techreport{mcnp,
+  address = {Los Alamos, NM, USA},
[…]
