+++ docs/paper.md
@@ -109,16 +109,18 @@ It uses the Numba compiler for Python to compile compute kernels to a desired ha
 These acceleration and abstraction techniques allow `MC/DC` developers to remain in a pure Python development environment without needing to support compiled or domain-specific languages.
 This has allowed `MC/DC` to grow from its initialization less than two years ago into a codebase that supports full performant neutron transport and investigation of novel transport algorithms, with development mostly from relative novices.
 
-Many of the traditionally developed neutron-transport codes are export-controlled (i.e., are not open source and difficult to access) and notoriously difficult to install, use, and develop in.
-Because `MC/DC` is an open-source and easily installable Python package (with a `pip`-installable distribution), it is ideal for use in an academic environment for both research and education.
+Many traditionally developed neutron-transport codes (e.g. `MCNP` [@mcnp], `Shift` [@shift], and `MCATK` [@mcatk]) are export-controlled and notoriously difficult to install, use, and develop in.
+`MC/DC` is comparable to other open source Monte Carlo neutron-transport apps (e.g. `OpenMC` [@openmc]), however `MC/DC` is an easily installable Python package (with a `pip` distribution) with it's own novel transport methods and features tailored to tackle dynamic transport.
+This all together makes `MC/DC` ideal for use in an academic environment for both research and education.
 This is further assisted by the test suite we have developed for unit, regression, verification, and performance tests, which are mostly run using continuous integration via GitHub Actions.
+In a code-to-code performance comparison `MC/DC` was found to run about 2.5 times slower then the Shift Monte Carlo code, and showed similar scaling on some systems [@variansyah_mc23_mcdc].
 
 `MC/DC` has support for continuous energy and multi-group treatments of the neutron distribution in energy.
 It can solve k-eigenvalue problems (used to determine neutron population growth rates in reactors) as well as fully dynamic simulations.
 It has a novel continuous geometry movement function that models transient elements (e.g., control rods or pulsed neutron experiments) more accurately than the step functions used by other codes.
 It also supports some simple domain decomposition, with more complex algorithms currently being implemented.
 
-`MC/DC`-enabled explorations into dynamic neutron transport algorithms have been successful, including quasi-Monte Carlo techniques [@mcdc:variansyah_physor22_pct], hybrid iterative techniques for k-eigenvalue simulations [@mcdc:qmc; @mcdc:qmcabs], transient population control techniques [@mcdc:variansyah_nse22_pct], hash-based random number generation, uncertainty and global sensitivity analysis [@mcdc:clements_mc23], residual Monte Carlo methods, and machine learning techniques for dynamic node scheduling, among others.
+`MC/DC`-enabled explorations into dynamic neutron transport algorithms have been successful, including quasi-Monte Carlo techniques [@mcdc:variansyah_physor22_pct], hybrid iterative techniques for k-eigenvalue simulations [@mcdc:qmc; @mcdc:qmcabs], transient population control techniques [@mcdc:variansyah_nse22_pct], hash-based random number generation, uncertainty and global sensitivity analysis [@mcdc:clements_mc23; mcdc:clements_variance_2024], residual Monte Carlo methods, and machine learning techniques for dynamic node scheduling, among others.
 
 # Future Work
 
