### 2024-03-26 CEMeNT-PSAAP/MCDC#174 @jpmorgan98
Major Points

- Agreed, the fix to this is in the draft PR, waiting on some of my collages to add descriptions for functions I am not sure about

Minor points

- Link added
- Instructions changed to match
- Many of these examples are provided to show a demonstration of an input deck for a given system. We modeled our input system to be similar to other production codes (`MCNP` and `OpenMC`) so we expect it to make more sense to folks who are familiar with those systems. Sometime soon we plan on expanding our documentation to include descriptions that you recommended. I have added some notes in some of the more complex input decks for recommended particle counts. This is also the reason for low particle counts in highly complex problems as you found in #173
- Fixed!
