### 2024-03-26 50503b4
adding citation to paper and notes to examples
