### 2024-03-26 2e5eab9
adding citation to paper and notes to examples
