+++ docs/paper.bib
@@ -82,6 +82,21 @@
     doi     = {10.1080/00295639.2022.2091906}
 }
 
+@article{mcdc:clements_variance_2024,
+	title = {A variance deconvolution estimator for efficient uncertainty quantification in {Monte} {Carlo} radiation transport applications},
+	volume = {319},
+	issn = {0022-4073},
+	url = {https://www.sciencedirect.com/science/article/pii/S0022407324000657},
+	doi = {https://doi.org/10.1016/j.jqsrt.2024.108958},
+	abstract = {Monte Carlo simulations are at the heart of many high-fidelity simulations and analyses for radiation transport systems. As is the case with any complex computational model, it is important to propagate sources of input uncertainty and characterize how they affect model output. Unfortunately, uncertainty quantification (UQ) is made difficult by the stochastic variability that Monte Carlo transport solvers introduce. The standard method to avoid corrupting the UQ statistics with the transport solver noise is to increase the number of particle histories, resulting in very high computational costs. In this contribution, we propose and analyze a sampling estimator based on the law of total variance to compute UQ variance even in the presence of residual noise from Monte Carlo transport calculations. We rigorously derive the statistical properties of the new variance estimator, compare its performance to that of the standard method, and demonstrate its use on neutral particle transport model problems involving both attenuation and scattering physics. We illustrate, both analytically and numerically, the estimator’s statistical performance as a function of available computational budget and the distribution of that budget between UQ samples and particle histories. We show analytically and corroborate numerically that the new estimator is unbiased, unlike the standard approach, and is more accurate and precise than the standard estimator for the same computational budget.},
+	journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
+	author = {Clements, Kayla B. and Geraci, Gianluca and Olson, Aaron J. and Palmer, Todd S.},
+	year = {2024},
+	keywords = {Monte Carlo radiation transport, Stochastic solvers, Uncertainty quantification},
+	pages = {108958},
+}
+}
+
 @inproceedings{variansyah_mc23_mcdc,
     Booktitle = {International Conference on Mathematics and Computational Methods Applied to Nuclear Science and Engineering},
     title = {Development of {MC/DC}: a performant, scalable, and portable {P}ython-based {M}onte {C}arlo neutron transport code},
@@ -196,4 +211,15 @@
     doi = {10.1016/S0168-9002(03)01368-8},
     url = {https://www.sciencedirect.com/science/article/pii/S0168900203013688},
[…]
