### 2024-03-27 c621ffa
Update paper.md
Fix citations
