+++ docs/paper.bib
@@ -65,12 +65,29 @@
 
 @inproceedings{mcdc:variansyah_physor22_pct,
     Booktitle = {International Conference on Physics of Reactors},
-    title = {An effective initial particle sampling technique for {M}onte {C}arlo reactor transient simulations},
+    title = {Performance of Population Control Techniques in Monte Carlo Reactor Criticality Simulation},
     year = {2022},
     author = {Ilham Variansyah and Ryan G. McClarren},
     address = {Pittsburgh, Pennsylvania, USA},
 }
 
+@inproceedings{variansyah_mc23_ic,
+    Booktitle = {International Conference on Mathematics and Computational Methods Applied to Nuclear Science and Engineering},
+    title = {An effective initial particle sampling technique for Monte Carlo reactor transient simulations},
+    year = {2023},
+    author = {Ilham Variansyah and Ryan G. McClarren},
+    address = {Niagara Falls, Ontario, Canada},
+    doi = {10.48550/arXiv.2305.07646}
+}
+
+@inproceedings{variansyah_mc23_moving_object,
+    Booktitle = {International Conference on Mathematics and Computational Methods Applied to Nuclear Science and Engineering},
+    title = {High-fidelity treatment for object movement in time-dependent Monte Carlo transport simulations},
+    year = {2023},
+    author = {Ilham Variansyah and Ryan G. McClarren},
+    address = {Niagara Falls, Ontario, Canada},
+    doi = {10.48550/arXiv.2305.07641}
+}
 
 @article{mcdc:variansyah_nse22_pct,
     author  = {Ilham Variansyah and Ryan G McClarren}, 
@@ -95,13 +112,12 @@
 	keywords = {Monte Carlo radiation transport, Stochastic solvers, Uncertainty quantification},
 	pages = {108958},
 }
-}
 
 @inproceedings{variansyah_mc23_mcdc,
     Booktitle = {International Conference on Mathematics and Computational Methods Applied to Nuclear Science and Engineering},
     title = {Development of {MC/DC}: a performant, scalable, and portable {P}ython-based {M}onte {C}arlo neutron transport code},
     year = {2023},
-    author = {Ilham Variansyah and Joanna Piper Morgan and Kyle E. Niemeyer and Ryan G. McClarren},
+    author = {Ilham Variansyah and Joanna Piper Morgan and Jordan Northrop and Kyle E. Niemeyer and Ryan G. McClarren},
     address = {Niagara Falls, Ontario, Canada},
     doi = {10.48550/arXiv.2305.07636}
 }
