+++ mcdc/input_.py
@@ -1,7 +1,7 @@
 """
 This module contains functions for setting MC/DC input deck.
 The input deck class is defined in `card.py` and instantiated in `global_.py`.
-Note docstring are numpy typed 
+Docstrings use NumPy formatting.
 """
 
 import h5py, math, mpi4py, os
@@ -62,7 +62,7 @@ def nuclide(
     Parameters
     ----------
     capture : numpy.ndarray (1D), optional
-        Capture microscopic cross-section [barn]
+        Capture microscopic cross-section [barn].
     scatter : numpy.ndarray (2D), optional
         Differential scattering microscopic cross-section [gout, gin] [barn].
     fission : numpy.ndarray (1D), optional
@@ -74,23 +74,23 @@ def nuclide(
     nu_d : numpy.ndarray (2D), optional
         Delayed neutron precursor yield [dg, gin].
     chi_p : numpy.ndarray (2D), optional
-        Prompt fission spectrum [gout, gin]
+        Prompt fission spectrum [gout, gin].
     chi_d : numpy.ndarray (2D), optional
-        Delayed neutron spectrum [gout, dg]
+        Delayed neutron spectrum [gout, dg].
     speed : numpy.ndarray (1D), optional
-        Energy group speed [cm/s]
+        Energy group speed [cm/s].
     decay : numpy.ndarray (1D), optional
-        Precursor group decay constant [/s]
+        Precursor group decay constant [/s].
     sensitivity : bool, optional
-        Set to `True` to calculate sensitivities to the nuclide
+        Set to `True` to calculate sensitivities to the nuclide.
     dsm_Np : float
         Average number of derivative particles produced at each
-        sensitivity nuclide collision
+        sensitivity nuclide collision.
 
     Returns
     -------
     dictionary
-        A nuclide card
+        A nuclide card.
 
     Notes
     -----
@@ -245,9 +245,9 @@ def material(
     Parameters
     ----------
     nuclides : list of tuple of (dictionary, float), optional
-        List of pairs of nuclide card and its density [/barn-cm]
+        List of pairs of nuclide card and its density [/barn-cm].
     capture : numpy.ndarray (1D), optional
-        Capture macroscopic cross-section [/cm]
+        Capture macroscopic cross-section [/cm].
     scatter : numpy.ndarray (2D), optional
         Differential scattering macroscopic cross-section [gout, gin] [/cm].
     fission : numpy.ndarray (1D), optional
@@ -259,24 +259,24 @@ def material(
     nu_d : numpy.ndarray (2D), optional
         Delayed neutron precursor yield [dg, gin].
     chi_p : numpy.ndarray (2D), optional
-        Prompt fission spectrum [gout, gin]
+        Prompt fission spectrum [gout, gin].
     chi_d : numpy.ndarray (2D), optional
-        Delayed neutron spectrum [gout, dg]
+        Delayed neutron spectrum [gout, dg].
     speed : numpy.ndarray (1D), optional
-        Energy group speed [cm/s]
+        Energy group speed [cm/s].
     decay : numpy.ndarray (1D), optional
-        Precursor group decay constant [/s]
+        Precursor group decay constant [/s].
     sensitivity : bool, optional
         Set to `True` to calculate sensitivities to the material
-        (only relevant for single-nuclide material)
+        (only relevant for single-nuclide material).
     dsm_Np : float
         Average number of derivative particles produced at each
-        sensitivity material collision (only relevant for single_nuclide material)
+        sensitivity material collision (only relevant for single_nuclide material).
 
     Returns
     -------
     dictionary
-        A material card
+        A material card.
 
     See also
     --------
[…]
