+++ docs/paper.bib
@@ -65,7 +65,7 @@
 
 @inproceedings{mcdc:variansyah_physor22_pct,
     Booktitle = {International Conference on Physics of Reactors},
-    title = {Performance of Population Control Techniques in Monte Carlo Reactor Criticality Simulation},
+    title = {Performance of Population Control Techniques in {M}onte {C}arlo Reactor Criticality Simulation},
     year = {2022},
     author = {Ilham Variansyah and Ryan G. McClarren},
     address = {Pittsburgh, Pennsylvania, USA},
@@ -73,7 +73,7 @@
 
 @inproceedings{variansyah_mc23_ic,
     Booktitle = {International Conference on Mathematics and Computational Methods Applied to Nuclear Science and Engineering},
-    title = {An effective initial particle sampling technique for Monte Carlo reactor transient simulations},
+    title = {An effective initial particle sampling technique for {M}onte {C}arlo reactor transient simulations},
     year = {2023},
     author = {Ilham Variansyah and Ryan G. McClarren},
     address = {Niagara Falls, Ontario, Canada},
@@ -82,7 +82,7 @@
 
 @inproceedings{variansyah_mc23_moving_object,
     Booktitle = {International Conference on Mathematics and Computational Methods Applied to Nuclear Science and Engineering},
-    title = {High-fidelity treatment for object movement in time-dependent Monte Carlo transport simulations},
+    title = {High-fidelity treatment for object movement in time-dependent {M}onte {C}arlo transport simulations},
     year = {2023},
     author = {Ilham Variansyah and Ryan G. McClarren},
     address = {Niagara Falls, Ontario, Canada},
@@ -124,7 +124,7 @@
 
 @article{brax2023,
     author = {Cuneo, Braxton and Bailey, Mike},
-    title = {Divergence Reduction in Monte Carlo Neutron Transport with On-GPU Asynchronous Scheduling},
+    title = {Divergence Reduction in {M}onte {C}arlo Neutron Transport with On-{GPU} Asynchronous Scheduling},
     year = {2023},
     publisher = {Association for Computing Machinery},
     address = {New York, NY, USA},
@@ -209,7 +209,7 @@
     organization = {UT-Battelle LLC/ORNL, Oak Ridge, TN (Unted States)},
     pages = {274-281},
     publisher = {Elsevier Ltd},
-    title = {The OpenMC Monte Carlo particle transport code},
+    title = {The {OpenMC} {M}onte {C}arlo particle transport code},
     volume = {51},
     year = {2013},
 }
@@ -230,7 +230,7 @@
 }
 
 @confrence{mcdc:cuneo2024alternative,
-      title={An Alternative to Stride-Based RNG for Monte Carlo Transport}, 
+      title={An Alternative to Stride-Based RNG for {M}onte {C}arlo Transport}, 
       author={Braxton S. Cuneo and Ilham Variansyah},
       year={2024},
       doi = {10.48550/arXiv.2403.06362},
