### 2024-03-27 #6415 @jpmorgan98
@lewisfish We have edited those sections of the paper to include comments about other software in this space and fixed/explained all the issues you raised in https://github.com/CEMeNT-PSAAP/MCDC/issues/173 and https://github.com/CEMeNT-PSAAP/MCDC/issues/174. The changes have been merged into our main branch. Let me know if there is anything we need from us to complete your review! Thank you again!
