### 2024-03-28 #6415 @lewisfish
> @lewisfish We have edited those sections of the paper to include comments about other software in this space and fixed/explained all the issues you raised in [CEMeNT-PSAAP/MCDC#173](https://github.com/CEMeNT-PSAAP/MCDC/issues/173) and [CEMeNT-PSAAP/MCDC#174](https://github.com/CEMeNT-PSAAP/MCDC/issues/174). The changes have been merged into our main branch. Let me know if there is anything we need from us to complete your review! Thank you again!

Looks great! One final issue I've noticed is that with the updated doc strings, it seems the readthedocs website has not been rebuilt with them. If you update the online docs I think that the review done from my side.
