### 2024-03-28 0643119
Added colorama mock for sphinx
