### 2024-03-28 4a8f083
Add colorama mock for Sphinx
