### 2024-04-02 #6415 @szabo137
@jpmorgan98 First of all, I like the package, and the paper, especially the focus on usability is very positive for scalable software, which is supposed to be executed in HPC environments.

Regarding this review, I have no concerns with this one.

Beyond this, you could think of adding a docker file to improve the installation procedure even further.
