### 2024-04-05 e668b50
adding inputs for caching
