### 2024-04-05 f951d6a
adding inputs for caching
