### 2024-04-08 #6415 @kellyrowland
@jpmorgan98 it looks like there are a few minor issues with references and citations in the paper:

- please add a DOI entry of `10.13182/physor22-37871` for the `mcdc:variansyah_physor22_pct` bibliography entry
- in line 121 of the paper, it looks like `[@mcdc:variansyah_mc23_mcdc]` should be `[variansyah_mc23_mcdc]` (or change the bibliograpy entry to match, whichever you like)
- line 123, `[@mcdc:variansyah_mc23_moving_object]` should be `[variansyah_mc23_moving_object]`
- line 123, `[@mcdc:variansyah_mc23_ic]` should be `[variansyah_mc23_ic]`

looks good otherwise! after those are fixed and the rendering looks good I'll post the next steps.
