### 2024-04-08 cc8772b
adding dois, fixing in text bib refs
