### 2024-04-08 42ab4a7
updating pyproject for version release
