+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "mcdc"
-version = "0.9.0"
+version = "0.9.1"
 authors = [
   { name="Ilham Variansyah", email="variansi@oregonstate.edu" },
   { name="Sam Pasmann", email="spasmann@nd.edu" },
@@ -14,8 +14,25 @@ authors = [
   { name="Charles Goodman"},
   { name="Caleb Shaw"},
   { name="Rohan Pankaj"},
-  { name="Jordan Northrop"},
+  { name="Alexander Mote"},
+  { name="Ethan Lame"},
+  { name="Benjamin Whewell"},
+  { name="Ryan G. McClarren"},
+  { name="Todd S. Palmer"},
+  { name="Lizhong Chen"},
+  { name="Dmitriy Y. Anistratov"},
+  { name="C. T. Kelley"},
+  { name="Camille J. Palmer"},
+  { name="Kyle E. Niemeyer"}
 ]
+
+maintainers = [
+    "Ilham Variansyah <variansi@oregonstate.edu>",
+    "Kayla Clements <clemekay@oregonstate.edu>",
+    "Joanna Piper Morgan <morgajoa@oregonstate.edu>",
+    "Kyle E. Niemeyer <kyle.niemeyer@oregonstate.edu>",
+]
+
 description = "Monte Carlo / Dynamic Code, a pure python high performance Monte Carlo neutronics package"
 readme = "README.md"
 requires-python = ">=3.9, <=3.11.8"
