### 2024-04-08 #6415 @jpmorgan98
Version number 0.9.1 [tagged](https://github.com/CEMeNT-PSAAP/MCDC/releases/tag/v0.9.1)
Zenodo Archive: [10.5281/zenodo.10576604](https://zenodo.org/doi/10.5281/zenodo.10576604)

Thanks
