### 2024-04-08 #6415 @kellyrowland
thanks @jpmorgan98 - can you edit the archive metadata so that:

- the licenses match (repo has BSD, archive lists CC)
- the authors are listed in the same order between the paper and the archive
