### 2024-04-12 #6415 @kyleniemeyer
@openjournals/dev it looks like @kellyrowland's name in the "Editor:" field has a different link color - and in fact the link there seems broken. Any idea what's going on? The ORCID link works ok. I assume that is supposed to link to a personal webpage?
